requirements.txt
selenium
pyyaml
aiohttp
mkdocs
mkdocs-minify-plugin
pymdown-extensions